feat: 자동 로또를 생성하는 객체 생성
