+++ src/main/kotlin/lotto/model/LottoMachine.kt
@@ -1,15 +1,7 @@
 package lotto.model
 
-import lotto.model.Lotto.Companion.LOTTO_NUMBER_MAX_RANGE
-import lotto.model.Lotto.Companion.LOTTO_NUMBER_MIN_RANGE
-import lotto.model.Lotto.Companion.LOTTO_NUMBER_SIZE
-
-class LottoMachine : LottoGenerator {
-    override fun getLottoNumbers(): Lotto {
-        val shuffledLottoNumbers = (LOTTO_NUMBER_MIN_RANGE..LOTTO_NUMBER_MAX_RANGE).shuffled()
-        val selectedLottoNumbers = shuffledLottoNumbers.take(LOTTO_NUMBER_SIZE).sorted()
-        val selectedLotto = selectedLottoNumbers.map { LottoNumber(it) }.toSet()
-
-        return Lotto(selectedLotto)
-    }
+class LottoMachine(
+    private val lottoGenerator: LottoGenerator,
+) {
+    fun generateLotto(): Lotto = lottoGenerator.getLottoNumbers()
 }
